remove crap from riak.lua
